GlobalAncova(GlobalAncova)	R Documentation

Global test for differential gene expression

Description

Computation of the sum of squares decomposition of gene-expression values according to a two way layout with gene and group as factors. Gene specific covariate effects are allowed. A MC approximation to a permutation-test for group main effect and gene: group interaction is provided. If just one gene should be tested a squared t-statistic which is equivalent to a F-statistic is computed. Corresponding p-values and permutation p-values are provided.

Usage


GlobalAncova(xx, group, covars = NULL, perm = 10000, test.genes = NULL)

Arguments

xx 	 A matrix of gene expression data, where columns correspond to samples and rows to genes. The data should be properly normalized beforehand (and log- or otherwise transformed). Missing values are not allowed. Gene and sample names can be included as the row and column names of xx.

group 	 A vector with the group membership information. In the given version group must be coded as 0-1.

covars 	 A vector or matrix which contains for each sample the covariate information.

perm 	 The number of permutations to be used. The default is 10,000.

test.genes 	 Vector of genes that shall be tested or list of pathways, each containing gene names.

Value

An ANOVA table

Note

This work was supported by the NGFN project 01 GR 0459, BMBF, Germany.

Author(s)

Reinhard Meister meister@tfh-berlin.de
Ulrich Mansmann mansmann@ibe.med.uni-muenchen.de

References

Mansmann, U. and Meister, R., 2005, Testing differential gene expression in functional groups, Methods Inf Med 44 (3).

See Also

Plot.genes, Plot.subjects, GlobalAncova.closed

Examples


set.seed(123)
data(p53.signalling)
data(cov.info)
data(group.info)
table.1   <- GlobalAncova(xx=p53.signalling, group=group.info, covars=NULL, perm=10000, test.genes = NULL)
table.2   <- GlobalAncova(xx=p53.signalling, group=group.info, covars=cov.info, perm=10000, test.genes = NULL)
table.sex <- GlobalAncova(xx=p53.signalling, group=cov.info[,1], covars=NULL, perm=10000, test.genes = NULL)
table.loc <- GlobalAncova(xx=p53.signalling, group=cov.info[,2], covars=NULL, perm=10000, test.genes = NULL)

[Package GlobalAncova version 1.2 Index]