GlobalAncova.1gene(GlobalAncova)	R Documentation

Test for differential gene expression for just one gene

Description

A squared t-statistic which is equivalent to a F-statistic is computed. Corresponding p-values and permutation p-values are provided.

Usage


GlobalAncova.1gene(xx, group, perm = 10000)

Arguments

xx 	 A vector of gene expression data. The data should be properly normalized beforehand (and log- or otherwise transformed). Missing values are not allowed.

group 	 A vector with the group membership information. In the given version group must be coded as 0-1.

perm 	 The number of permutations to be used. The default is 10,000.

Value

A vector giving the F value, permutation p–value and theoretical p–value.

Note

For internal use.

Author(s)

Reinhard Meister meister@tfh-berlin.de
Ulrich Mansmann mansmann@ibe.med.uni-muenchen.de
Manuela Hummel hummel@ibe.med.uni-muenchen.de

See Also

GlobalAncova, GlobalAncova.closed, Plot.genes, Plot.subjects

[Package GlobalAncova version 1.2 Index]