GlobalAncova.closed(GlobalAncova)	R Documentation

Closed testing procedure for testing several pathways using GlobalAncova

Description

Computation of a closed testing procedure for several pathways as an alternative of correcting for multiple testing. Starting from the pathways of interest a family of null hypotheses is created that is closed under intersection. Each null hypothesis can be rejected at a given level if it is rejected along with all hypotheses included in it. For details see Marcus, R., Peritz, E. and Gabriel, K. R., 1976, On closed testing procedures with special reference to ordered analysis of variance, Biometrika 63 (3): 655–660.

Usage


GlobalAncova.closed(xx, group, covars = NULL, test.genes, previous.test = NULL, level = 0.05, perm = 10000)

Arguments

xx 	 A matrix of gene expression data, where columns correspond to samples and rows to genes. The data should be properly normalized beforehand (and log- or otherwise transformed). Missing values are not allowed. Gene and sample names can be included as the row and column names of xx.

group 	 A vector with the group membership information. In the given version group must be coded as 0-1.

covars 	 A vector or matrix which contains for each sample the covariate information.

test.genes 	 A list of pathways that shall be tested, each containing gene names.

previous.test 	 The output of a call to 'GlobalAncova' with specified option 'test.genes' according to the pathways of interest.

level 	 The level of significance of the testing procedure.

perm 	 The number of permutations to be used. The default is 10,000.

Value

A list with components

new.data 	 Family of null hypotheses (vectors of genes to be tested simultaneously with 'GlobalAncova').

test.results 	 Test results for each pathway of interest and all hypotheses included in it.

significant 	 Names of the significant pathways.

not.significant 	 Names of the not significant pathways.

Note

This work was supported by the NGFN project 01 GR 0459, BMBF, Germany.

Author(s)

Reinhard Meister meister@tfh-berlin.de
Ulrich Mansmann mansmann@ibe.med.uni-muenchen.de
Manuela Hummel hummel@ibe.med.uni-muenchen.de

See Also

GlobalAncova, Plot.genes, Plot.subjects

[Package GlobalAncova version 1.2 Index]