Hnull.family(GlobalAncova)	R Documentation

Family of null hypotheses closed under intersection

Description

Creates new groups of genes by building unions of the given pathways. Each group of genes corresponds to the null hypothesis that those genes are not differentially expressed. All groups together build a family of null hypotheses that is closed under intersection.

Usage


Hnull.family(test.genes)

Arguments

test.genes 	 A list of pathways that shall be tested, each containing gene names.

Value

A list of groups of genes each corresponding to one hypothesis.

Note

For internal use.

Author(s)

Reinhard Meister meister@tfh-berlin.de
Ulrich Mansmann mansmann@ibe.med.uni-muenchen.de
Manuela Hummel hummel@ibe.med.uni-muenchen.de

References

Marcus, R., Peritz, E. and Gabriel, K. R., 1976, On closed testing procedures with special reference to ordered analysis of variance, Biometrika 63 (3): 655–660.

See Also

GlobalAncova, GlobalAncova.closed, Plot.genes, Plot.subjects

[Package GlobalAncova version 1.2 Index]