Plot.subjects(GlobalAncova)	R Documentation

Subjects Plot for Global Ancova

Description

Produces a plot to show the influence of the samples on the test result produced by 'GlobalAncova'.

Usage


Plot.subjects(xx, group, covars = NULL, sort = FALSE)

Arguments

xx 	 A matrix of gene expression data, where columns correspond to samples and rows to genes. The data should be properly normalized beforehand (and log- or otherwise transformed). Missing values are not allowed. Gene and sample names can be included as the row and column names of xx.

group 	 A vector with the group membership information. In the given version group must be coded as 0-1.

covars 	 A vector or matrix which contains for each sample the covariate information.

sort 	 Should the samples be ordered by group?

Note

This work was supported by the NGFN project 01 GR 0459, BMBF, Germany.

Author(s)

Reinhard Meister meister@tfh-berlin.de
Ulrich Mansmann mansmann@ibe.med.uni-muenchen.de

See Also

GlobalAncova, Plot.genes, GlobalAncova.closed

Examples


data(p53.signalling)
data(group.info)

Plot.subjects(xx=p53.signalling, group=group.info, covars=NULL, sort=FALSE)

[Package GlobalAncova version 1.2 Index]