horizontal.bars(GlobalAncova)	R Documentation

Plotting horizontal bars

Description

Function for plotting horizontal bars which is needed in 'Plot.genes' and 'Plot.subjects'.

Usage


horizontal.bars(x, xlabel = "", ylabel = "", color = NULL, labelsize = 0.75, bar.names = NULL)

Arguments

x 	 A vector of values to be plot.

xlabel 	 A title for the x axis.

ylabel 	 A title for the y axis.

color 	 Colours of the bars.

labelsize 	 The magnification to be used for axis annotation relative to the current.

bar.names 	 Labels of the bars.

Note

For internal use.

Author(s)

Reinhard Meister meister@tfh-berlin.de
Ulrich Mansmann mansmann@ibe.med.uni-muenchen.de

See Also

Plot.genes, Plot.subjects, GlobalAncova, GlobalAncova.closed

[Package GlobalAncova version 1.2 Index]